+++ line.h
@@ -7,12 +7,12 @@ class Line
 {
 public:
     std::string text;   // text of this line
-    size_t len;         // length of this line (including CRLF)
-    size_t x;           // top left corner
-    size_t y;
-    size_t w;           // width
-    size_t h;           // height
-    size_t base;        // base line
+    int len;         // length of this line (including CRLF)
+    int x;           // top left corner
+    int y;
+    int w;           // width
+    int h;           // height
+    int base;        // base line
     Line * prev;
     Line * next;
 };
